+++ twine/commands/register.py
@@ -24,8 +24,12 @@ from twine import utils
 
 
 def register(package, repository, username, password, comment, config_file,
-             cert, client_cert):
-    config = utils.get_repository_from_config(config_file, repository)
+             cert, client_cert, repository_url):
+    config = utils.get_repository_from_config(
+        config_file,
+        repository,
+        repository_url,
+    )
     config["repository"] = utils.normalize_repository_url(
         config["repository"]
     )
@@ -63,15 +67,26 @@ def main(args):
     parser.add_argument(
         "-r", "--repository",
         action=utils.EnvironmentDefault,
-        env='TWINE_REPOSITORY',
+        env="TWINE_REPOSITORY",
         default="pypi",
-        help="The repository to register the package to (default: "
+        help="The repository to register the package to. Can be a section in "
+             "the config file or a full URL to the repository (default: "
              "%(default)s)",
+    )
+    parser.add_argument(
+        "--repository-url",
+        action=utils.EnvironmentDefault,
+        env="TWINE_REPOSITORY_URL",
+        default=None,
+        required=False,
+        help="The repository URL to upload the package to. This can be "
+             "specified with --repository because it will be used if there is "
+             "no configuration for the value passed to --repository."
     )
     parser.add_argument(
         "-u", "--username",
         action=utils.EnvironmentDefault,
-        env='TWINE_USERNAME',
+        env="TWINE_USERNAME",
         required=False, help="The username to authenticate to the repository "
                              "as (can also be set via %(env)s environment "
                              "variable)",
@@ -79,7 +94,7 @@ def main(args):
     parser.add_argument(
         "-p", "--password",
         action=utils.EnvironmentDefault,
-        env='TWINE_PASSWORD',
+        env="TWINE_PASSWORD",
         required=False, help="The password to authenticate to the repository "
                              "with (can also be set via %(env)s environment "
                              "variable)",