+++ twine/utils.py
@@ -40,6 +40,7 @@ else:
 
 
 DEFAULT_REPOSITORY = "https://upload.pypi.org/legacy/"
+TEST_REPOSITORY = "https://test.pypi.org/legacy/"
 
 
 def get_config(path="~/.pypirc"):
@@ -50,7 +51,11 @@ def get_config(path="~/.pypirc"):
         return {"pypi": {"repository": DEFAULT_REPOSITORY,
                          "username": None,
                          "password": None
-                         }
+                         },
+                "pypitest": {"repository": TEST_REPOSITORY,
+                             "username": None,
+                             "password": None
+                             },
                 }
 
     # Parse the rc file
@@ -94,13 +99,22 @@ def get_config(path="~/.pypirc"):
     return config
 
 
-def get_repository_from_config(config_file, repository):
+def get_repository_from_config(config_file, repository, repository_url=None):
     # Get our config from the .pypirc file
     try:
         return get_config(config_file)[repository]
     except KeyError:
+        if repository_url and "://" in repository_url:
+            # assume that the repsoitory is actually an URL and just sent
+            # them a dummy with the repo set
+            return {
+                "repository": repository_url,
+                "username": None,
+                "password": None,
+            }
         msg = (
-            "Missing '{repo}' section from the configuration file.\n"
+            "Missing '{repo}' section from the configuration file\n"
+            "or not a complete URL in --repository.\n"
             "Maybe you have a out-dated '{cfg}' format?\n"
             "more info: "
             "https://docs.python.org/distutils/packageindex.html#pypirc\n"